Optimize codegen for If, Block and Loop
If we know statically that a code block only has one exit (local or non-local) we can elide the runtime-checks.

Be aware: the implementation here is a bit naive, and can be sped up to produce the same code, but faster.

In my benchmarks, this saves about 8% of code in the bitcoin header verifier.  From 43669 lines to 40603 lines of MASM.